cursor_assignment[HW1]/requirements.txt
requests==2.32.3
schedule==1.2.2
python-telegram-bot==13.15
urllib3<2
six==1.16.0




